requirements_win.txt
asttokens==2.4.1
cffi==1.16.0
colorama==0.4.6
comm==0.2.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest_asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pickleshare==0.7.5
pip==24.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.9.0
pytz==2024.1
pywin32==305.1
pyzmq==25.1.2
rpy2==3.5.11
setuptools==69.1.1
simplegeneric==0.8.1
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
wcwidth==0.2.13
wheel==0.42.0
zipp==3.17.0
